Fix missing cast in paged attention test
In the flurry of PRs that went in yesterday,
a casting bug was introduced in the paged attention test.
This PR fixes that by introducing an explicit cast
to f16 inside the phase one kernel.

+++ lit_tests/kernel/wave/attention.py
@@ -1121,7 +1121,7 @@ def test_paged_flash_decoding():
     v = torch.randn(v_shape, dtype=torch.float16)
     logits = torch.zeros(logits_shape, dtype=torch.float32)
     logits_max = torch.zeros(logits_max_shape, dtype=torch.float32)
-    output = torch.zeros(o_shape, dtype=torch.float32)
+    output = torch.zeros(o_shape, dtype=torch.float16)
 
     with tk.gen.TestLaunchContext(
         hyperparams_0,
